requirements.txt
flask==3.1.0
pillow==11.1.0
google-cloud-vision==3.4.4
google-generativeai==0.8.4
langchain==0.3.15
langchain-google-genai==2.0.9
langchain-community==0.3.15
faiss-cpu==1.9.0
PyPDF2==3.0.1
python-dotenv==1.0.1
Werkzeug==3.1.3
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1


